weights

+++ src/data/LPR_dataset.py
@@ -90,7 +90,7 @@ def collate_fn(batch):
         images.append(torch.from_numpy(img))
         labels.extend(label)
         lengths.append(length)
-    labels = np.asarray(labels).flatten().astype(np.float32)
+    labels = np.asarray(labels).flatten()
     return (torch.stack(images, 0), torch.from_numpy(labels), lengths)
